whatsappcrm_backend/requirements.txt
# Django Core
django
drf-nested-routers
# Environment variable management
python-dotenv

# Admin interface theme
django-jazzmin

# REST API framework
djangorestframework
djangorestframework-simplejwt

# CORS handling
django-cors-headers

# Celery and related packages
celery
django-celery-results
django-celery-beat
redis # For Celery broker if using Redis

# Database connector (PostgreSQL)
psycopg2-binary # Or psycopg2 if you manage build dependencies

# WSGI server for production
gunicorn

# ASGI server (since ASGI_APPLICATION is defined)
daphne
requests
pydantic
# WebSocket support
 # For Redis channel layer
# For handling file uploads
django-storages
# For Additional Functionality
django-filter # For filtering in DRF
django-extensions # For additional management commands
drf-spectacular # For OpenAPI schema generation
django-prometheus # For monitoring
Pillow
jinja2
paynow
channels
channels_redis
openpyxl
reportlab
django-csp